requirements.txt
pandas==2.1.1
plotly-express==0.4.1
streamlit==1.27.2